% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1v3p.R
\name{model1v3p}
\alias{model1v3p}
\title{Modeling function 1v3p}
\usage{
model1v3p(model.data, ny, nx, CV = F, CV_n = 1, r2random = F,
  runs = 1000)
}
\arguments{
\item{model.data=data.frame}{Data.frame that contains both the dependent and independent variables}

\item{ny=number}{Number of dependent variables to be tested.}

\item{nx=number}{Number of independent variables to be tested. Do not confuse number of available independent
variables with number of independent variables OF THE MODEL.}

\item{CV=boolean}{T if a cross-validation should be performed, F if not.}

\item{CV_n=number}{Number of observations to be left out in the cross-validation process. For example
CV = T & CV_n = 1 will perform a 1-leave-out-cross-validation, while the same command with CV_n = 2, will
performa a 2-leave-out-cross-validation}

\item{r2random=boolean}{T if an R^2 maximum random distribution should be computed with the data. Calculating
this random distribution enables a comparison of the observed R^2 values of the best models with a completely
random scenario. This distribution shows if the goodness-of-fit values obtained in the models correspond to a
significntly higher value than the expected at random (>= value of percentile 95) or not.}

\item{runs=number}{Number indicating the number of runs to perform the goodness-of-fit random distribution.}
}
\description{
This function constructs every possible linear model with one independent variable with an
exponential term(y = mx^2 + b) for nx number of dependent variables and testing every independent variable (nx).
}
\examples{
model1v3p(df,ny=9,nx=16)

}
